adding jwt claims for admin privileges

+++ resources/recipe.py
@@ -2,7 +2,7 @@ import uuid
 from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 
 from schemas import RecipeSchema
 
@@ -23,6 +23,10 @@ class Recipe(MethodView):
     
     @jwt_required()
     def delete(self, recipe_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+        
         recipe = RecipeModel.query.get_or_404(recipe_id)
         db.session.delete(recipe)
         db.session.commit()
@@ -38,6 +42,10 @@ class RecipeList(MethodView):
     @blp.arguments(RecipeSchema)
     @blp.response(201, RecipeSchema)
     def post(self, request_recipe):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+
         recipe = RecipeModel(**request_recipe)
         try:
             db.session.add(recipe)
